RP10T40 Refactoring package persistance to entities, configs to configurations
